webapp/requirements.txt
Flask == 2.3.*
mysql-connector-python == 8.0.*
argon2-cffi == 21.3.*
gunicorn == 21.*
passlib == 1.7.*


